requirements.txt
symphony-bdk-python>=2.10.0
Jinja2~=3.0
